Fixed an issue with metadata on load and save of cells
